% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv.gspca}
\alias{cv.gspca}
\title{CV for generalized supervised PCA}
\usage{
cv.gspca(x, y, ks, alphas = 10^seq(-3, 5, by = 0.5), family_x, family_y,
  folds = 5, quiet = TRUE, ...)
}
\arguments{
\item{x}{covariate matrix}

\item{y}{response vector}

\item{ks}{the different dimensions \code{k} to try}

\item{alphas}{the different approximations to the saturated model \code{m} to try}

\item{family_x}{exponential family distribution of covariates}

\item{family_y}{exponential family distribution of response}

\item{folds}{if \code{folds} is a scalar, then it is the number of folds. If
it is a vector, it should be the same length as the number of rows in \code{x}}

\item{quiet}{logical; whether the function should display progress}

\item{...}{Additional arguments passed to \code{logisticPCA}}
}
\value{
A matrix of the CV deviance for the response with \code{k} in rows and
 \code{alpha} in columns
}
\description{
Run cross validation on dimension and \code{alpha} for generalized supervised PCA
}
\examples{
# construct a low rank matrix in the logit scale
rows = 100
cols = 10
set.seed(1)
mat_logit = outer(rnorm(rows), rnorm(cols))

# generate a binary matrix
mat = (matrix(runif(rows * cols), rows, cols) <= inv.logit.mat(mat_logit)) * 1.0

\dontrun{
negloglikes = cv.gspca(mat, ks = 1:9)
plot(negloglikes)
}
}

