% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sup_mf.R
\name{genSupMF}
\alias{genSupMF}
\title{Supervised matrix factorization for exponential family data}
\usage{
genSupMF(x, y, k = 2, alpha = NULL, family_x = c("gaussian", "binomial",
  "poisson"), family_y = c("gaussian", "binomial", "poisson"), quiet = TRUE,
  max_iters = 1000, conv_criteria = 1e-05, random_start = FALSE, start_A,
  start_B, mu)
}
\arguments{
\item{x}{covariate matrix}

\item{y}{response vector}

\item{k}{dimension}

\item{alpha}{balance between dimensionality reduction of \code{x} and prediction of \code{y}}

\item{family_x}{exponential family distribution of covariates}

\item{family_y}{exponential family distribution of response}

\item{quiet}{logical; whether the calculation should give feedback}

\item{max_iters}{maximum number of iterations}

\item{conv_criteria}{convergence criteria}

\item{random_start}{whether to randomly initialize \code{A} and \code{B}}

\item{start_A}{initial value for \code{A}}

\item{start_B}{initial value for \code{B}}

\item{mu}{specific value for \code{mu}, the mean vector of \code{x}}
}
\value{
An S3 object of class \code{gsmf} which is a list with the
following components:
\item{mu}{the main effects for dimensionality reduction}
\item{A}{the \code{n}x\code{k}-dimentional matrix with the scores}
\item{B}{the \code{d}x\code{k}-dimentional matrix with the loadings}
\item{beta}{the \code{k + 1} length vector of the coefficients}
\item{family_x}{the exponential family of covariates}
\item{family_y}{the exponential family of response}
\item{iters}{number of iterations required for convergence}
\item{loss_trace}{the trace of the average negative log likelihood of the algorithm.
   Should be non-increasing}
}
\description{
Supervised matrix factorization for exponential family data
}
\examples{
rows = 100
cols = 10
set.seed(1)
mat_np = outer(rnorm(rows), rnorm(cols))

# generate a count matrix and binary response
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)
response = rbinom(rows, 1, rowSums(mat) / max(rowSums(mat)))

mod = genSupMF(mat, response, k = 1, alpha = 1, family_x = "poisson", family_y = "binomial",
               quiet = FALSE)

plot(inv.logit.mat(cbind(1, mod$A) \%*\% mod$beta), response)
plot(rowSums(mat), response)
\dontrun{
ggplot(data.frame(PC = mod$PCs[, 1], y = response), aes(PC, y)) + stat_summary_bin(bins = 10)
}
}
\references{
Rish, Irina, et al. "Closed-form supervised dimensionality
  reduction with generalized linear models." Proceedings of the 25th
  international conference on Machine learning. ACM, 2008.
}

