% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{plot.cv.gspca}
\alias{plot.cv.gspca}
\title{Plot CV for generalized supervised PCA}
\usage{
\method{plot}{cv.gspca}(x, ...)
}
\arguments{
\item{x}{a \code{cv.gspca} object}

\item{...}{Additional arguments}
}
\description{
Plot cross validation results generalized supervised PCA
}
\examples{
# construct a low rank matrix in the natural parameter space
rows = 100
cols = 10
set.seed(1)
mat_np = outer(rnorm(rows), rnorm(cols))

# generate a count matrix
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)

\dontrun{
loglikes = cv.gpca(mat, ks = 1:9, Ms = 3:6, family = "poisson")
plot(loglikes)
}
}

