% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sup_mf.R
\name{predict.gsmf}
\alias{predict.gsmf}
\title{Predict response with a generalized supervised MF model}
\usage{
\method{predict}{gsmf}(object, newdata, type = c("link", "response", "PCs"),
  quiet = TRUE, max_iters = 1000, conv_criteria = 1e-05, start_A, ...)
}
\arguments{
\item{object}{generalized supervised MF object}

\item{newdata}{matrix of the same exponential family as covariates in \code{object}.
If missing, will use the data that \code{object} was fit on}

\item{type}{the type of fitting required.
\code{type = "link"} gives response variable on the natural parameter scale,
\code{type = "response"} gives response variable on the response scale, and
\code{type = "PCs"} gives matrix of principal components of \code{x}}

\item{quiet}{logical; whether the calculation should show progress}

\item{max_iters}{maximum number of iterations}

\item{conv_criteria}{convergence criteria}

\item{start_A}{initial value for \code{A}}

\item{...}{Additional arguments}
}
\description{
Predict response with a generalized supervised MF model
}
\examples{
# construct a low rank matrices in the natural parameter space
rows = 100
cols = 10
set.seed(1)
loadings = rnorm(cols)
mat_np = outer(rnorm(rows), rnorm(cols))
mat_np_new = outer(rnorm(rows), loadings)

# generate a count matrices and binary responses
mat = matrix(rbinom(rows * cols, 1, c(inv.logit.mat(mat_np))), rows, cols)
mat_new = matrix(rbinom(rows * cols, 1, c(inv.logit.mat(mat_np_new))), rows, cols)

response = rbinom(rows, 1, rowSums(mat) / max(rowSums(mat)))
response_new = rbinom(rows, 1, rowSums(mat_new) / max(rowSums(mat_new)))

mod = genSupMF(mat, response, k = 2, alpha = 1000,
               family_x = "poisson", family_y = "binomial", quiet = FALSE)

plot(predict(mod, type = "response"), response)
plot(predict(mod, mat_new, type = "response"), response_new)

}

