% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sup_pca.R
\name{predict.gspca}
\alias{predict.gspca}
\title{Predict response with a generalized supervised PCA model}
\usage{
\method{predict}{gspca}(object, newdata, type = c("link", "response"), ...)
}
\arguments{
\item{object}{generalized supervised PCA object}

\item{newdata}{matrix of the same exponential family as covariates in \code{object}.
If missing, will use the data that \code{object} was fit on}

\item{type}{the type of fitting required.
\code{type = "link"} gives matrix on the natural parameter scale and
\code{type = "response"} gives matrix on the response scale}

\item{...}{Additional arguments}
}
\examples{
# construct a low rank matrices in the natural parameter space
rows = 100
cols = 10
set.seed(1)
loadings = rnorm(cols)
mat_np = outer(rnorm(rows), rnorm(cols))
mat_np_new = outer(rnorm(rows), loadings)

# generate a count matrices and binary responses
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)
mat_new = matrix(rpois(rows * cols, c(exp(mat_np_new))), rows, cols)

response = rbinom(rows, 1, rowSums(mat) / max(rowSums(mat)))
response_new = rbinom(rows, 1, rowSums(mat_new) / max(rowSums(mat_new)))

mod = genSupPCA(mat, response, k = 1, family_x = "poisson", family_y = "binomial",
                quiet = FALSE, max_iters_per = 1, discrete_deriv = FALSE)

plot(predict(mod, mat, type = "response"), response_new)

}

